# api/openapi.yaml
openapi: 3.0.3
info:
  title: Route Beacon API
  version: 2.0.0
  description: |
    Read-only REST API for the BGP Looking Glass. Serves route data from a
    PostgreSQL database populated by the external BMP ingester (route-beacon-ri).

    ## Architecture

    This API is read-only. The database is written exclusively by the
    route-beacon-ri ingester service. The API queries `routers`,
    `current_routes`, `route_events`, `rib_sync_status`, and the
    `route_summary` materialized view.

    ## Error Format

    All error responses use RFC 7807 Problem Details (`application/problem+json`).

  contact:
    name: Route Beacon
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:8080
    description: Local development

paths:
  # --------------------------------------------------------------------------
  # Health
  # --------------------------------------------------------------------------
  /api/v1/health:
    get:
      operationId: getHealth
      summary: System health check
      description: |
        Returns the overall health of the API and a summary of monitored routers.
        Health state is derived from router sync status and database connectivity.
      tags: [health]
      responses:
        "200":
          description: System is healthy or degraded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
              example:
                status: healthy
                router_count: 3
                online_routers: 3
                total_routes: 850000
                uptime_seconds: 86400
        "503":
          description: System is unhealthy (database unreachable or no routers).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"

  # --------------------------------------------------------------------------
  # Routers
  # --------------------------------------------------------------------------
  /api/v1/routers:
    get:
      operationId: listRouters
      summary: List all monitored routers
      description: |
        Returns every router known to the system with its metadata and
        online/offline status derived from BMP session state.
      tags: [routers]
      responses:
        "200":
          description: Router list.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Router"
              example:
                data:
                  - id: "10.0.0.2"
                    router_ip: "172.28.0.10"
                    hostname: "bgp-router1"
                    display_name: "bgp-router1"
                    as_number: 65001
                    description: "ISP Alpha core router"
                    status: "up"
                    eor_received: true
                    first_seen: "2026-02-20T10:00:00Z"
                    last_seen: "2026-02-21T14:30:00Z"
        "500":
          $ref: "#/components/responses/InternalError"

  /api/v1/routers/{routerId}:
    get:
      operationId: getRouter
      summary: Get a single router
      description: Returns full details for one monitored router.
      tags: [routers]
      parameters:
        - $ref: "#/components/parameters/RouterId"
      responses:
        "200":
          description: Router details.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/Router"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # --------------------------------------------------------------------------
  # Route Lookup
  # --------------------------------------------------------------------------
  /api/v1/routers/{routerId}/routes/lookup:
    get:
      operationId: lookupRoutes
      summary: Look up routes for a prefix
      description: |
        Returns all BGP paths the selected router holds for the given prefix.

        **Match behaviour**: when `match_type` is omitted, the server auto-detects
        based on input. A CIDR prefix (e.g. `8.8.8.0/24`) triggers an exact match;
        a bare IP address (e.g. `8.8.8.8`) triggers a longest-prefix match.
      tags: [routes]
      parameters:
        - $ref: "#/components/parameters/RouterId"
        - name: prefix
          in: query
          required: true
          description: |
            IPv4 or IPv6 prefix in CIDR notation (e.g. `8.8.8.0/24`,
            `2001:db8::/32`) or a bare IP address for longest-prefix match.
          schema:
            type: string
          examples:
            cidr_v4:
              summary: IPv4 CIDR prefix (exact match)
              value: "8.8.8.0/24"
            bare_ip_v4:
              summary: Bare IPv4 address (longest-prefix match)
              value: "8.8.8.8"
            cidr_v6:
              summary: IPv6 CIDR prefix (exact match)
              value: "2001:db8::/32"
        - name: match_type
          in: query
          required: false
          description: |
            Force a specific match type. When omitted the server auto-detects
            based on whether the input includes a prefix length.
          schema:
            type: string
            enum: [exact, longest]
      responses:
        "200":
          description: Route lookup results.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteLookupResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"

  # --------------------------------------------------------------------------
  # Route History (API-only, no frontend UI)
  # --------------------------------------------------------------------------
  /api/v1/routers/{routerId}/routes/history:
    get:
      operationId: getRouteHistory
      summary: Route change history for a prefix
      description: |
        Returns historical route events (announcements and withdrawals) for a
        given prefix on the selected router. Results are ordered by time
        descending (most recent first).

        The `route_events` table is partitioned by day on `ingest_time`. Always
        include time bounds for efficient queries. Default range is last 24 hours.
      tags: [routes]
      parameters:
        - $ref: "#/components/parameters/RouterId"
        - name: prefix
          in: query
          required: true
          description: CIDR prefix to look up history for (e.g. `10.100.0.0/24`).
          schema:
            type: string
        - name: from
          in: query
          required: false
          description: Start time (ISO 8601). Defaults to 24 hours ago.
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: false
          description: End time (ISO 8601). Defaults to now.
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          required: false
          description: Maximum number of events to return. Default 100, max 1000.
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
      responses:
        "200":
          description: Route history events.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteHistoryResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"

# ==========================================================================
# Components
# ==========================================================================
components:
  # --------------------------------------------------------------------------
  # Parameters
  # --------------------------------------------------------------------------
  parameters:
    RouterId:
      name: routerId
      in: path
      required: true
      description: Router identifier (matches `routers.router_id` in the database).
      schema:
        type: string
      example: "10.0.0.2"

  # --------------------------------------------------------------------------
  # Schemas
  # --------------------------------------------------------------------------
  schemas:
    # -- Health --------------------------------------------------------------
    HealthResponse:
      type: object
      required:
        - status
        - router_count
        - online_routers
        - total_routes
        - uptime_seconds
      properties:
        status:
          type: string
          enum: [healthy, degraded, unhealthy]
          description: |
            Overall system health. "healthy" = all online routers have received
            End-of-RIB. "degraded" = some routers still syncing. "unhealthy" =
            no online routers or database unreachable.
        router_count:
          type: integer
          description: Total number of known routers.
        online_routers:
          type: integer
          description: Routers with active BMP sessions.
        total_routes:
          type: integer
          description: Total routes across all routers (from materialized view).
        uptime_seconds:
          type: integer
          description: Seconds since the API server started.

    # -- Router --------------------------------------------------------------
    Router:
      type: object
      required:
        - id
        - display_name
        - status
        - eor_received
        - first_seen
        - last_seen
      properties:
        id:
          type: string
          description: Router identifier (BGP Router ID).
        router_ip:
          type: string
          nullable: true
          description: IP address the router connected from.
        hostname:
          type: string
          nullable: true
          description: Router hostname from BMP sysName.
        display_name:
          type: string
          description: Human-friendly name (hostname or router_id).
        as_number:
          type: integer
          nullable: true
          description: Router's autonomous system number.
        description:
          type: string
          nullable: true
          description: Router description from BMP sysDescr.
        status:
          type: string
          enum: [up, down]
          description: |
            "up" if the router has an active BMP session (rib_sync_status row
            exists), "down" if the session has terminated.
        eor_received:
          type: boolean
          description: Whether End-of-RIB has been received for this router.
        first_seen:
          type: string
          format: date-time
          description: When this router was first observed.
        last_seen:
          type: string
          format: date-time
          description: When the last BMP session initiation was received.

    # -- Route ---------------------------------------------------------------
    Community:
      type: object
      required:
        - type
        - value
      properties:
        type:
          type: string
          enum: [standard, extended, large]
        value:
          type: string
          description: |
            Community string value. Standard: "ASN:value". Extended: "RT:ASN:value"
            or "SOO:ASN:value". Large: "GA:LD1:LD2".

    Route:
      type: object
      required:
        - prefix
        - path_id
        - as_path
        - origin
        - communities
        - extended_communities
        - large_communities
        - first_seen
        - updated_at
      properties:
        prefix:
          type: string
          description: Network prefix in CIDR notation.
        path_id:
          type: integer
          description: BGP Add-Path identifier. 0 when Add-Path is not in use.
        next_hop:
          type: string
          nullable: true
          description: BGP next-hop address.
        as_path:
          description: |
            AS path as an array. Each element is either an integer (AS number) or
            an array of integers (AS_SET).
          oneOf:
            - type: array
              items:
                type: integer
            - type: array
              items:
                oneOf:
                  - type: integer
                  - type: array
                    items:
                      type: integer
        origin:
          type: string
          enum: [igp, egp, incomplete]
          nullable: true
          description: BGP origin attribute (lowercased).
        local_pref:
          type: integer
          nullable: true
          description: LOCAL_PREF value.
        med:
          type: integer
          nullable: true
          description: Multi-Exit Discriminator.
        origin_asn:
          type: integer
          nullable: true
          description: Last ASN in as_path (the origin AS). Null if path is empty or ends with AS_SET.
        communities:
          type: array
          items:
            $ref: "#/components/schemas/Community"
          description: Standard BGP communities.
        extended_communities:
          type: array
          items:
            $ref: "#/components/schemas/Community"
          description: Extended BGP communities.
        large_communities:
          type: array
          items:
            $ref: "#/components/schemas/Community"
          description: Large BGP communities.
        attrs:
          type: object
          nullable: true
          description: Extra BGP path attributes not mapped to dedicated fields.
        first_seen:
          type: string
          format: date-time
          description: When this route was first inserted.
        updated_at:
          type: string
          format: date-time
          description: Last time this route was updated.

    # -- Route Lookup Response -----------------------------------------------
    RouterSummary:
      type: object
      required:
        - id
        - display_name
        - as_number
      properties:
        id:
          type: string
        display_name:
          type: string
        as_number:
          type: integer
          nullable: true

    RouteLookupMeta:
      type: object
      required:
        - match_type
        - router_status
      properties:
        match_type:
          type: string
          enum: [exact, longest]
          description: The match type that was applied.
        router_status:
          type: string
          enum: [up, down]
          description: Current BMP session status of the router.

    RouteLookupResponse:
      type: object
      required:
        - prefix
        - router
        - routes
        - meta
      properties:
        prefix:
          type: string
          description: The matched prefix in CIDR notation.
        router:
          $ref: "#/components/schemas/RouterSummary"
        routes:
          type: array
          items:
            $ref: "#/components/schemas/Route"
          description: All routes for the matched prefix on this router.
        plain_text:
          type: string
          description: Pre-formatted plain-text rendering for copy-to-clipboard.
        meta:
          $ref: "#/components/schemas/RouteLookupMeta"

    # -- Route History Response ----------------------------------------------
    RouteEvent:
      type: object
      required:
        - timestamp
        - action
        - prefix
      properties:
        timestamp:
          type: string
          format: date-time
          description: When the event was recorded (ingest_time).
        action:
          type: string
          enum: [announce, withdraw]
          description: '"announce" for route add/update, "withdraw" for removal.'
        prefix:
          type: string
          description: Affected prefix in CIDR notation.
        path_id:
          type: integer
          nullable: true
          description: Add-Path identifier. Null when not applicable.
        next_hop:
          type: string
          nullable: true
          description: Next-hop for announcements. Null for withdrawals.
        as_path:
          type: array
          nullable: true
          items:
            oneOf:
              - type: integer
              - type: array
                items:
                  type: integer
          description: AS path at time of event. Null for withdrawals.
        origin:
          type: string
          nullable: true
        local_pref:
          type: integer
          nullable: true
        med:
          type: integer
          nullable: true
        origin_asn:
          type: integer
          nullable: true
        communities:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/Community"
        extended_communities:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/Community"
        large_communities:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/Community"

    RouteHistoryResponse:
      type: object
      required:
        - router_id
        - prefix
        - events
      properties:
        router_id:
          type: string
        prefix:
          type: string
        from:
          type: string
          format: date-time
          description: Start of the time range queried.
        to:
          type: string
          format: date-time
          description: End of the time range queried.
        events:
          type: array
          items:
            $ref: "#/components/schemas/RouteEvent"
        has_more:
          type: boolean
          description: True if more events exist beyond the limit.

    # -- Error Responses (RFC 7807) ------------------------------------------
    ProblemDetail:
      type: object
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
          default: "about:blank"
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri
        invalid_params:
          type: array
          items:
            type: object
            required:
              - name
              - reason
            properties:
              name:
                type: string
              reason:
                type: string

  # --------------------------------------------------------------------------
  # Reusable Responses
  # --------------------------------------------------------------------------
  responses:
    BadRequest:
      description: The request is malformed.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"

    NotFound:
      description: The requested resource does not exist.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
          example:
            type: "about:blank"
            title: Not Found
            status: 404
            detail: "Router '10.0.0.99' does not exist."

    ValidationError:
      description: One or more request parameters are invalid.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
          example:
            type: "about:blank"
            title: Unprocessable Entity
            status: 422
            detail: "Request validation failed."
            invalid_params:
              - name: prefix
                reason: "Not a valid IPv4 or IPv6 prefix."

    InternalError:
      description: An unexpected server error occurred.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
          example:
            type: "about:blank"
            title: Internal Server Error
            status: 500
            detail: "An unexpected error occurred."

tags:
  - name: health
    description: System health and status.
  - name: routers
    description: Monitored BGP router listing and details.
  - name: routes
    description: BGP route lookup and history.